// src/entities/Article/model/consts/consts.ts
export enum ArticleSortField {
    VIEWS = 'views',
    TITTLE = 'tittle',
    CREATED = 'createdAt',
}

export enum ArticleBlockType {
    CODE = 'CODE',
    TEXT = 'TEXT',
    IMAGE = 'IMAGE'
}

export enum ArticleType {
    ALL = 'ALL',
    IT = 'IT',
    SCIENCE = 'SCIENCE',
    ECONOMICS = 'ECONOMICS',
}

export enum ArticleView {
    BIG = 'big',
    SMALL = 'small',
}